fix bug

+++ Source/Services/Roles/Role.cs
@@ -206,6 +206,12 @@ namespace Insight.Base.Services
         /// <returns>bool 是否成功</returns>
         public bool Delete()
         {
+            if (BuiltIn)
+            {
+                Result.NotBeDeleted();
+                return false;
+            }
+
             var result = DbHelper.Delete(_Role);
             if (!result) Result.DataBaseError();
